UPSTREAM: zram: support page writeback
There is demand to writeback specific process pages to backing store
instead of all idles pages in the system due to storage wear out concerns
and to launching latency of apps which are most of the time idle but are
critical for resume latency.

This patch extends the writeback knob to support a specific page
writeback.

Link: https://lkml.kernel.org/r/20201020190506.3758660-1-minchan@kernel.org
Signed-off-by: Minchan Kim <minchan@kernel.org>
Reviewed-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit 0d8359620d9be9823b6b9b3cf2dbe006cbfec594)

Bug: 181035934
Signed-off-by: Amos Bianchi <amosbianchi@google.com>
Change-Id: I2c2ef973f66c9d780244d39a1833c4246fb28bc2
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>