drm/msm/sde: Skip unneeded register reads when getting write line count
More often than not, get_vsync_info() is used to only get the write line
count, while the other values it returns are left unused. This is not
optimal since it is done on every display commit. We can eliminate the
superfluous register reads by adding a parameter specifying if only the
write line count is requested.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Yuan Si <do4suki@gmail.com>
Change-Id: I80e5922c5a24a7eaef99b74e30185be2e14b5e0d
Signed-off-by: fadlyas07 <mhmmdfdlyas@proton.me>